[level 1] Title: 정수 제곱근 판별, Time: 0.03 ms, Memory: 77.7 MB -BaekjoonHub

+++ src/programmers/lv1/lv1_12934/Solution.java
@@ -2,6 +2,6 @@ package programmers.lv1.lv1_12934;
 
 class Solution {
     public long solution(long n) {
-        return Math.pow((int)Math.sqrt(n), 2) == n ? (long)Math.pow((int)Math.sqrt(n)+1, 2) : -1;
+        return Math.pow((int)Math.sqrt(n), 2) == n ? (long)Math.pow(Math.sqrt(n)+1, 2) : -1;
     }
 }
